+++ hash_functions.h
@@ -8,7 +8,7 @@
 
 uint32_t hash_SuperFastHash(const char *data, int len); // these might need to be unsigned char
 
-uint32_t hash_djb2(const char *str);
+unsigned long hash_djb2(const char *str);
 
 uint32_t hash_jenkins_one_at_a_time(const char *data, int len);
